Revert "perf(interpreter): improve ATC interprater's cache efficiency (#246)"
This reverts commit 6fd34eb19ab4d16f012707d49df93f3f3e91782f.
