+++ src/router.rs
@@ -1,4 +1,4 @@
-use crate::cir::{CirProgram, Translate};
+use crate::ast::Expression;
 use crate::context::{Context, Match};
 use crate::interpreter::Execute;
 use crate::parser::parse;
@@ -12,7 +12,7 @@ struct MatcherKey(usize, Uuid);
 
 pub struct Router<'a> {
     schema: &'a Schema,
-    matchers: BTreeMap<MatcherKey, CirProgram>,
+    matchers: BTreeMap<MatcherKey, Expression>,
     pub fields: HashMap<String, usize>,
 }
 
@@ -33,10 +33,11 @@ impl<'a> Router<'a> {
         }
 
         let ast = parse(atc).map_err(|e| e.to_string())?;
+
         ast.validate(self.schema)?;
-        let cir = ast.translate();
-        cir.add_to_counter(&mut self.fields);
-        assert!(self.matchers.insert(key, cir).is_none());
+        ast.add_to_counter(&mut self.fields);
+
+        assert!(self.matchers.insert(key, ast).is_none());
 
         Ok(())
     }
@@ -44,8 +45,8 @@ impl<'a> Router<'a> {
     pub fn remove_matcher(&mut self, priority: usize, uuid: Uuid) -> bool {
         let key = MatcherKey(priority, uuid);
 
-        if let Some(cir) = self.matchers.remove(&key) {
-            cir.remove_from_counter(&mut self.fields);
+        if let Some(ast) = self.matchers.remove(&key) {
+            ast.remove_from_counter(&mut self.fields);
             return true;
         }
